requirements.txt
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.7.2
FlagEmbedding==1.3.5
torch==2.9.1
transformers==4.57.1
openai==2.8.1
umap-learn==0.5.9.post2
hdbscan==0.8.40
pyod==2.0.5
qdrant-client==1.16.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.44
alembic==1.17.2
redis==7.1.0
rq==2.6.0
unstructured[pdf]==0.18.20
pdf2image==1.17.0
pytesseract==0.3.13
python-dotenv==1.2.1
tqdm==4.67.1
click==8.3.1
streamlit==1.51.0
plotly==6.5.0
watchgod==0.8.2
spacy==3.8.11
pypdf==6.3.0
python-docx==1.2.0
beautifulsoup4==4.13.5
lxml==5.4.0
reportlab==4.4.5
extract-msg==0.55.0
Pillow==12.0.0
